# lib/api-zod/src/generated/types/updateRoleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AppWorld API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateRoleInputRole } from './updateRoleInputRole';

export interface UpdateRoleInput {
  role: UpdateRoleInputRole;
}